remove act for userevents, create utility function to mock link creation and download instead
