+++ packages/synapse-react-client/src/components/CitationPopover/CitationPopoverUtils.ts
@@ -1,0 +1,13 @@
+export const createLinkAndDownload = (
+  title: string,
+  fileExtension: string,
+  url: string,
+) => {
+  const citationTitle = `${title.replace(/\s+/g, '_')}`
+  const link = document.createElement('a')
+  link.href = url
+  link.setAttribute('download', `${citationTitle}.${fileExtension}`)
+  document.body.appendChild(link)
+  link.click()
+  document.body.removeChild(link)
+}
